quantum_circuits.py
import numpy as np
from dwave.system import DWaveSampler, EmbeddingComposite
from collections import defaultdict

class QuantumCircuit:
    _sampler = None  # Class-level sampler instance

    @classmethod
    def get_sampler(cls):
        """Get or create the quantum sampler instance."""
        if cls._sampler is None:
            cls._sampler = EmbeddingComposite(DWaveSampler())
        return cls._sampler

    def __init__(self, num_qubits):
        """Initialize quantum circuit with specified number of qubits."""
        self.num_qubits = num_qubits
        self.sampler = self.get_sampler()

    def create_superposition_state(self):
        """Create superposition state using quantum sampler."""
        Q = {(i, i): -1 for i in range(self.num_qubits)}
        return self.sampler.sample_qubo(Q, num_reads=100)

    def apply_quantum_gate(self, gate_type, target_qubits):
        """Apply quantum gates using quantum sampler."""
        Q = {}

        if gate_type == "X":  # NOT gate
            for qubit in target_qubits:
                Q[(qubit, qubit)] = -1

        elif gate_type == "CNOT":  # Controlled-NOT gate
            control, target = target_qubits
            Q[(control, control)] = -1
            Q[(target, target)] = -1
            Q[(control, target)] = 2

        elif gate_type == "Hadamard":  # Hadamard-like operation
            for qubit in target_qubits:
                Q[(qubit, qubit)] = 0

        return self.sampler.sample_qubo(Q, num_reads=100)

    def measure_state(self, quantum_state):
        """Measure quantum state and return most probable outcome."""
        state_counts = defaultdict(int)
        for record in quantum_state.record:
            state = tuple(record.sample)
            state_counts[state] += record.num_occurrences

        most_common_state = max(state_counts.items(), key=lambda x: x[1])[0]
        return list(most_common_state)

def create_quantum_hash(data, num_qubits=8):
    """Create a deterministic quantum-enhanced hash."""
    circuit = QuantumCircuit(num_qubits)

    # Convert input data to a deterministic binary string
    if isinstance(data, str):
        binary_data = ''.join(format(ord(c), '08b') for c in data)
    else:
        binary_data = format(int.from_bytes(data, 'big'), '08b')

    # Initialize quantum state
    state = circuit.create_superposition_state()

    # Apply quantum gates deterministically based on input data
    for i, bit in enumerate(binary_data[:num_qubits]):
        if bit == '1':
            state = circuit.apply_quantum_gate("X", [i % num_qubits])

    # Apply final Hadamard-like operation for superposition
    state = circuit.apply_quantum_gate("Hadamard", range(num_qubits))

    # Measure and convert to hash deterministically
    measured_state = circuit.measure_state(state)
    hash_value = int(''.join(str(bit) for bit in measured_state), 2)

    return hash_value

def verify_quantum_signature(message, signature, num_qubits=8):
    """Verify a signature using quantum circuits."""
    circuit = QuantumCircuit(num_qubits)

    # Create quantum hash of message
    message_hash = create_quantum_hash(message, num_qubits)

    # Create quantum hash of signature
    signature_hash = create_quantum_hash(signature, num_qubits)

    # Compare hashes using quantum CNOT
    comparison_state = circuit.apply_quantum_gate(
        "CNOT", 
        [message_hash % num_qubits, signature_hash % num_qubits]
    )

    # Measure result
    result = circuit.measure_state(comparison_state)

    # If states match, most qubits should be in |0⟩ state
    return sum(result) < num_qubits // 2