pyproject.toml
[project]
name = "repl-nix-quantumoptimizer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dwave-ocean-sdk>=8.1.0",
    "dwave-preprocessing>=0.6.7",
    "eth-account>=0.13.4",
    "networkx>=3.4.2",
    "numpy>=2.2.1",
    "plotly>=5.24.1",
    "pycryptodome>=3.21.0",
    "streamlit>=1.41.1",
    "web3>=7.6.1",
]
